+++ app/Http/Controllers/Admin/Career/CoverLetterController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Admin\Career;
 
 use App\Http\Controllers\Admin\BaseAdminController;
+use App\Http\Requests\Career\StoreCoverLettersRequest;
 use App\Http\Requests\Career\UpdateCoverLettersRequest;
 use App\Models\Career\CoverLetter;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
 use Illuminate\View\View;
@@ -75,6 +77,8 @@ class CoverLetterController extends BaseAdminController
      */
     public function edit(CoverLetter $coverLetter): View
     {
+        Gate::authorize('update-resource', $coverLetter);
+
         return view('admin.career.cover-letter.edit', compact('coverLetter'));
     }
 
@@ -88,6 +92,8 @@ class CoverLetterController extends BaseAdminController
     public function update(UpdateCoverLettersRequest $updateCoverLettersRequest,
                            CoverLetter               $coverLetter): RedirectResponse
     {
+        Gate::authorize('update-resource', $coverLetter);
+
         $coverLetter->update($updateCoverLettersRequest->validated());
 
         return redirect()->route('admin.career.cover-letter.show', $coverLetter)
@@ -102,6 +108,8 @@ class CoverLetterController extends BaseAdminController
      */
     public function destroy(CoverLetter $coverLetter): RedirectResponse
     {
+        Gate::authorize('delete-resource', $coverLetter);
+
         $coverLetter->delete();
 
         return redirect(referer('admin.career.cover-letter.index'))
