+++ app/Http/Controllers/Admin/Personal/RecipeController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Personal\UpdateRecipesRequest;
 use App\Models\Personal\Recipe;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 /**
@@ -74,6 +75,8 @@ class RecipeController extends BaseAdminController
      */
     public function edit(Recipe $recipe): View
     {
+        Gate::authorize('update-resource', $recipe);
+
         return view('admin.personal.recipe.edit', compact('recipe'));
     }
 
@@ -86,6 +89,8 @@ class RecipeController extends BaseAdminController
      */
     public function update(UpdateRecipesRequest $updateRecipesRequest, Recipe $recipe): RedirectResponse
     {
+        Gate::authorize('update-resource', $recipe);
+
         $recipe->update($updateRecipesRequest->validated());
 
         return redirect()->route('admin.personal.recipe.show', $recipe)
@@ -100,6 +105,8 @@ class RecipeController extends BaseAdminController
      */
     public function destroy(Recipe $recipe): RedirectResponse
     {
+        Gate::authorize('delete-resource', $recipe);
+
         $recipe->delete();
 
         return redirect(referer('admin.personal.recipe.index'))
